Issue #38 - FtdUsbPort timeout values being cast to a byte.

+++ NModbus/src/FtdAdapter/FtdUsbPort.cs
@@ -160,7 +160,7 @@ namespace FtdAdapter
 				_writeTimeout = (uint) value;
 
 				if (IsOpen)
-					InvokeFtdMethod(() => FT_SetTimeouts(DeviceHandle, (byte) ReadTimeout, (byte) WriteTimeout));
+					InvokeFtdMethod(() => FT_SetTimeouts(DeviceHandle, (uint) ReadTimeout, (uint) WriteTimeout));
 			}
 		}
